- Added support for prepared statement in update change; refactored executable prepared statements
