+++ liquibase-core/src/main/java/liquibase/change/core/InsertDataChange.java
@@ -1,16 +1,5 @@
 package liquibase.change.core;
 
-import java.io.BufferedInputStream;
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
-import java.io.FileReader;
-import java.io.StringReader;
-import java.math.BigDecimal;
-import java.math.BigInteger;
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -21,10 +10,8 @@ import liquibase.change.ChangeProperty;
 import liquibase.change.ChangeWithColumns;
 import liquibase.change.ColumnConfig;
 import liquibase.database.Database;
-import liquibase.database.PreparedStatementFactory;
 import liquibase.database.core.InformixDatabase;
-import liquibase.exception.DatabaseException;
-import liquibase.statement.ExecutablePreparedStatement;
+import liquibase.statement.InsertExecutablePreparedStatement;
 import liquibase.statement.SqlStatement;
 import liquibase.statement.core.InsertStatement;
 
@@ -103,7 +90,9 @@ public class InsertDataChange extends AbstractChange implements ChangeWithColumn
         }
 
         if (needsPreparedStatement) {
-            return new SqlStatement[]{ new InsertExecutablePreparedStatement(database) };
+            return new SqlStatement[] { 
+            		new InsertExecutablePreparedStatement(database, catalogName, schemaName, tableName, columns) 
+            };
         }
 
 
@@ -129,113 +118,5 @@ public class InsertDataChange extends AbstractChange implements ChangeWithColumn
      */
     public String getConfirmationMessage() {
         return "New row inserted into " + getTableName();
-    }
-
-
-    /**
-     * Handles INSERT Execution
-     */
-    public class InsertExecutablePreparedStatement implements ExecutablePreparedStatement {
-
-        private Database database;
-
-        InsertExecutablePreparedStatement(Database database) {
-            this.database = database;
-        }
-
-        public void execute(PreparedStatementFactory factory) throws DatabaseException {
-            // build the sql statement
-            StringBuilder sql = new StringBuilder("INSERT INTO ");
-            StringBuilder params = new StringBuilder("VALUES(");
-            sql.append(database.escapeTableName(getCatalogName(), getSchemaName(), getTableName()));
-            sql.append("(");
-            List<ColumnConfig> cols = new ArrayList<ColumnConfig>(getColumns().size());
-            for(ColumnConfig column : getColumns()) {
-                if(database.supportsAutoIncrement()
-                    && Boolean.TRUE.equals(column.isAutoIncrement())) {
-                    continue;
-                }
-                sql.append(database.escapeColumnName(getCatalogName(), getSchemaName(), getTableName(), column.getName()));
-                sql.append(", ");
-                params.append("?, ");
-                cols.add(column);
-            }
-            sql.deleteCharAt(sql.lastIndexOf(" "));
-            sql.deleteCharAt(sql.lastIndexOf(","));
-            params.deleteCharAt(params.lastIndexOf(" "));
-            params.deleteCharAt(params.lastIndexOf(","));
-            params.append(")");
-            sql.append(") ");
-            sql.append(params);
-
-            // create prepared statement
-            PreparedStatement stmt = factory.create(sql.toString());
-
-            try {
-                // attach params
-                int i = 1;  // index starts from 1
-                for(ColumnConfig col : cols) {
-                    if(col.getValue() != null) {
-                        stmt.setString(i, col.getValue());
-                    } else if(col.getValueBoolean() != null) {
-                        stmt.setBoolean(i, col.getValueBoolean());
-                    } else if(col.getValueNumeric() != null) {
-                        Number number = col.getValueNumeric();
-                        if(number instanceof Long) {
-                            stmt.setLong(i, number.longValue());
-                        } else if(number instanceof Integer) {
-                            stmt.setInt(i, number.intValue());
-                        } else if(number instanceof Double) {
-                            stmt.setDouble(i, number.doubleValue());
-                        } else if(number instanceof Float) {
-                            stmt.setFloat(i, number.floatValue());
-                        } else if(number instanceof BigDecimal) {
-                            stmt.setBigDecimal(i, (BigDecimal)number);
-                        } else if(number instanceof BigInteger) {
-                            stmt.setInt(i, number.intValue());
-                        }
-                    } else if(col.getValueDate() != null) {
-                        stmt.setDate(i, new java.sql.Date(col.getValueDate().getTime()));
-                    } else if(col.getValueBlob() != null) {
-                        try {
-                            File file = new File(col.getValueBlob());
-                            stmt.setBinaryStream(i, new BufferedInputStream(new FileInputStream(file)), (int) file.length());
-                        } catch (FileNotFoundException e) {
-                            throw new DatabaseException(e.getMessage(), e); // wrap
-                        }
-                    } else if(col.getValueClob() != null) {
-                        try {
-                            File file = new File(col.getValueClob());
-                            stmt.setCharacterStream(i, new BufferedReader(new FileReader(file)), (int) file.length());
-                        } catch(FileNotFoundException e) {
-                            throw new DatabaseException(e.getMessage(), e); // wrap
-                        }
-                    } else if(col.getValueText() != null) {
-                    	try {
-                    		StringReader r = new StringReader(col.getValueText());
-                    		stmt.setCharacterStream(i, new BufferedReader(r), col.getValueText().length());
-                    		//byte[] bytes = col.getValueText().getBytes("UTF-8");
-                    		//ByteArrayInputStream b = new ByteArrayInputStream(bytes);
-                    		//stmt.setBinaryStream(i, new BufferedInputStream(b), (int) col.getValueText().length());
-                    	} catch(Exception e) {
-                    		throw new DatabaseException(e.getMessage(), e);
-                    	}
-                    } else {
-                    	// NULL values might intentionally be set into a change, we must also add them to the prepared statement  
-                    	stmt.setNull(i, java.sql.Types.NULL);
-                    }
-                    i++;
-                }
-
-                // trigger execution
-                stmt.execute();
-            } catch(SQLException e) {
-                throw new DatabaseException(e);
-            }
-        }
-
-        public boolean skipOnUnsupported() {
-            return false;
-        }
     }
 }
