+++ liquibase-core/src/main/java/liquibase/sqlgenerator/core/InsertDataChangeGenerator.java
@@ -1,20 +1,20 @@
 package liquibase.sqlgenerator.core;
 
-import liquibase.change.core.InsertDataChange;
 import liquibase.database.Database;
 import liquibase.exception.ValidationErrors;
 import liquibase.sql.Sql;
 import liquibase.sqlgenerator.SqlGeneratorChain;
+import liquibase.statement.InsertExecutablePreparedStatement;
 
 /**
  * Dummy SQL generator for <code>InsertDataChange.ExecutableStatement</code><br>
  */
-public class InsertDataChangeGenerator extends AbstractSqlGenerator<InsertDataChange.InsertExecutablePreparedStatement> {
-    public ValidationErrors validate(InsertDataChange.InsertExecutablePreparedStatement statement, Database database, SqlGeneratorChain sqlGeneratorChain) {
+public class InsertDataChangeGenerator extends AbstractSqlGenerator<InsertExecutablePreparedStatement> {
+    public ValidationErrors validate(InsertExecutablePreparedStatement statement, Database database, SqlGeneratorChain sqlGeneratorChain) {
         return new ValidationErrors();
     }
 
-    public Sql[] generateSql(InsertDataChange.InsertExecutablePreparedStatement statement, Database database, SqlGeneratorChain sqlGeneratorChain) {
+    public Sql[] generateSql(InsertExecutablePreparedStatement statement, Database database, SqlGeneratorChain sqlGeneratorChain) {
         return new Sql[0];
     }
 }
